requirements.txt
# Production Requirements
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pandas>=2.0.0

# Security and monitoring
watchdog>=3.0.0
psutil>=5.9.0

# Optional: For better performance
torch>=2.0.0
transformers>=4.30.0